+++ noodles-bam/examples/bam_count_async.rs
@@ -6,17 +6,17 @@ use std::env;
 
 use futures::TryStreamExt;
 use noodles_bam as bam;
-use tokio::{fs::File, io};
+use tokio::fs::File;
 
 #[tokio::main]
-async fn main() -> io::Result<()> {
+async fn main() -> Result<(), Box<dyn std::error::Error>> {
     let src = env::args().nth(1).expect("missing src");
 
     let mut reader = File::open(src).await.map(bam::AsyncReader::new)?;
-    reader.read_header().await?;
+    let header = reader.read_header().await?.parse()?;
     reader.read_reference_sequences().await?;
 
-    let mut records = reader.records();
+    let mut records = reader.records(&header);
     let mut n = 0;
 
     while records.try_next().await?.is_some() {